# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/sessions/{sessionId}/end": {
      "post": {
        "operationId": "SessionController_endSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
    "/api/user/sessions/{sessionId}/progress": {
      "post": {
        "operationId": "SessionController_recordProgress",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
    "/api/user/sessions/history": {
      "get": {
        "operationId": "SessionController_getHistory",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/user/sessions/diff": {
      "get": {
        "operationId": "SessionController_diff",
        "parameters": [
          {
            "name": "dateA",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateB",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/v1/sessions/{id}/end": {
      "post": {
        "operationId": "SessionV1Controller_end",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SessionV1"
        ]
      }
    },
    "/api/v1/sessions/{id}/progress": {
      "post": {
        "operationId": "SessionV1Controller_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SessionV1"
        ]
      }
    },
    "/api/v1/sessions/history": {
      "get": {
        "operationId": "SessionV1Controller_history",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SessionV1"
        ]
      }
    },
...
    "/api/v1/sessions/diff": {
      "get": {
        "operationId": "SessionV1Controller_diff",
        "parameters": [
          {
            "name": "dateA",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateB",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SessionV1"
        ]
      }
    },
...
    "/api/user/resumes/{id}": {
      "get": {
        "operationId": "ResumeController_getResume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resume"
        ]
      }
    },
...
    "/api/user/resumes/{id}/branch": {
      "post": {
        "operationId": "ResumeController_createVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Resume"
        ]
      }
    },
...
    "/api/user/resumes/{id}/pdf": {
      "get": {
        "operationId": "ResumeController_exportPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resume"
        ]
      }
    },
    "/api/v1/resumes": {
      "get": {
        "operationId": "ResumeV1Controller_list",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      }
    },
    "/api/v1/resumes/{id}": {
      "get": {
        "operationId": "ResumeV1Controller_detail",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      },
      "put": {
        "operationId": "ResumeV1Controller_update",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      },
      "delete": {
        "operationId": "ResumeV1Controller_remove",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      }
    },
    "/api/v1/resumes/generate": {
      "post": {
        "operationId": "ResumeV1Controller_generate",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      }
    },
    "/api/v1/resumes/{id}/branch": {
      "post": {
        "operationId": "ResumeV1Controller_branch",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      }
    },
    "/api/v1/resumes/{id}/pdf": {
      "get": {
        "operationId": "ResumeV1Controller_pdf",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 resumes"
        ]
      }
    },
...
    "/api/user/quick-test": {
      "get": {
        "operationId": "QuickTestController_getQuestions",
        "parameters": [
          {
            "name": "direction",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuickTest"
        ]
      }
    },
...
  },
  "info": {
    "title": "ZhiPath Platform API",
    "description": "智途 ZhiPath、CodeNova 与后续定制客户端共享平台 API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "client-app": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-App"
      }
    },
...
  }
}